qcacld: Add status check before 'vdev_id' is used
The 'vdev_id' is assigned by calling
wmi_extract_roam_scan_stats_res_evt(), but acctualy the 'vdev_id'
may not be set in this function when it returns QDF_STATUS_E_FAILURE.
So have to check the return value of the function, then use 'vdev_id'.
This issue is reported from the Misra scanning.

Change-Id: I9b13b9f7b57e6ba7589ddf3d6fe838d78f9b8f6a