qca-wifi-oss: Validate if the ADFS chosen pdev is valid
When we do modeswitch from DBS to DBS_SBS mode and vice-versa with
preCAC config enabled, there is a chance that
dfs_find_pdev_for_agile_precac() may choose a pdev that is not
valid for the current hw mode. Hence, validate if the pdev
returned by dfs_find_pdev_for_agile_precac() is valid for the
current operating hw mode.

CRs-Fixed: 3664739
Change-Id: Icc8be91be5bf05ae7899b31e49a1052f54aef16f