qca-wifi-oss:  Traffic from Root to Repeater-2 is not going through
Issue: Traffic from Root to Repeater-2(STA only) when we configure
Root as DBDC, Repeater-1 as DBDC and Repeater-2 as
Single Band(non-primary)

Ucast pkt(ARP resp) is dropped at RE-1 in Secondary STA Intf in Rx.
Ideally it is suppose to be enqueued to the AP VAP of same band,
so that it will be forwarded to the single band(non-primary) STA.

Since STA's br-lan mac is learnt with AP_VLAN intf(athx.stax) in RE1,
the exisiting condition fails because it checks only for the AP intf(athx).
Added AP_VLAN intf also as an additional check.

Change-Id: Iabb2b36608f285a8eac86e516ceb5bd19028db8e
Signed-off-by: sivaraman <quic_sivad@quicinc.com>