qcacld-3.0: Move bss state check to hdd_ssr_restart_sap
In the multi-link SAP case, we will pend the cac link
and let the 6 GHz link will start firstly, which will
cause the bss has possibility to set to started state
in advance. So move the bss state checking outside
hdd_ssr_restart_sap() that can avoid the valid sap
content not to be free in the hdd_ssr_restart_sap
by mistake.

Change-Id: I7666266b52b3cb09352df5b8464c106984dad84d
CRs-Fixed: 3893945