qcacmn: Add wraper layer to check if peer exist on pdev
This is required for finding peer in any SOC with mac
addr and pdev id in case of roaming, client can connect
to another interface/repeater/node without sending any
disconnect/deauth/disassoc on current pdev/soc.

this is used in sending CLIENT_CONNECT_EX and CLIENT_DISCONNECT
event to IPA in case of roaming.

Change-Id: I61eb3ab09ff6fe5a07c29c61708ba5dcaa9e61ea