qcacld-3.0: Correct auth frame length if algorithm not support
As SAP mode, if received the authentication frame with
algorithm that it was not support, then should send with
status code 'STATUS_NOT_SUPPORTED_AUTH_ALG(0xd)' without
any additional bytes. But in currently code logic it will
carry challenge text length that is not right, so raise
change to correct the length for this case.

Change-Id: Ia63c52680391d5d5551817774ee4c8dde4d612df
CRs-Fixed: 3936042