qca-wifi: Add unpunctured CAC channels to NOL
When a channel is punctured due to radar detection, the radar-affected
channel is added to the NOL. Upon NOL timeout, the channel undergoes a
6-minute CAC, after which it is unpunctured. If radar is detected during
CAC, the CAC timer is reset, but the channel is not added back to NOL,
leading to a DFS violation.

For example, if an AP operates on channel 52 (80MHz) and radar is
detected on channel 64, channel 64 is added to NOL, and channel 52
becomes 80-20MHz. After NOL timeout, the AP performs a 6-minute CAC on
channel 64. If radar is detected during this CAC, the timer resets, but
channel 64 is not added back to NOL, causing a potential DFS violation.

To prevent this, if radar is detected on a channel that just came out of
NOL and started CAC, stop the CAC and add the channel back to NOL. Since
the channel has not changed, there is no need to compute a new
puncturing pattern.

Changes:
- During the CAC WAIT state, if a radar detected event is received,
  previously the CAC timer was just reset. Now, the CAC timer is
  cancelled, channels in the puncture SM are added back to NOL and the
  state moves to PUNCTURED.

Change-Id: Ia9e1f4e457013892b6df01b17daa9a6f8b5587dc
CRs-Fixed: 4056540 3920249
(cherry picked from commit 55d2787ea65cbb9fb585a5191e1c00701ed04326)