qcacmn: Free scan entry and force_connect_candidate
Currently force_connect_candidate is allocated but
in case of failure the candidate is not freed which
may cause issue.

While removing the scan entries which are added in
DLM list with reason code DLM_REMOVE, the entry is
removed from scan list but that entry is not freed.

Fix is to Free force_connect_candidate in case of
failure and free scan entry which is removed from
scan list.

Change-Id: Ieaed67bcbe57e2b3fa5349bbe0bd4df8f13605a1
CRs-Fixed: 3940861