qcacld-3.0: Revert "support of mgmt frame registration update ops"
This reverts change I87bcf44ac1601fe7beea4e5dd0320a68ac2c8748 as
it causes out of sync issues when ROC is just started and a frame
is received from fw in parallel. The frame can be sent to kernel
and kernel drops it if no userspace entity has registered for it.

Keep the registration/deregistration bitmap logs
which might be helpful for debugging.

Change-Id: I62fc05e501428dad88b596e9105111e0c7f779ec
CRs-Fixed: 3943266