qcacld-3.0: Fix cdp_vdev and dp_link cleanup
Currently dp_link handle is saved in dp_vdev as a part
of dp_vdev_register which happens upon connection.
Due to this, the dp_link cannot be freed until the
dp_vdev has been freed, else there are chances that
dp_link handle might be accessed even after its freed.

To make sure that dp_link is freed only after dp_vdev
is freed, there is a notification when dp_vdev is
supposed to be freed. This notification along with
objmgr_vdev destroy notification is used to free
the dp_link.

The above solution has a corner case where same dp_link
can get registered to multiple dp_vdevs and due to this
when one of the dp_vdev gets freed, the objmgr vdev
destroy notification does not wait for all the dp_vdevs
to be freed and goes ahead and frees the dp_link.

Scenario:
- DP-L0(id=0) created, DP-V0(id=0) created
- DP-L0(id=0) linked to DP-V0(id=0) during dp_vdev_register (STA
connect)
- STA disconnect occurs and subsequently mac address update is done
- Mac address update:
	- DP-V0(id=0) is detached, but not freed since there are
	  some reference release pending.
	- DP-V0'(id=0) is created with new mac address.
	- DP-L0(id=0) mac address is updated directly.
- DP-L0(id=0) is linked to DP-V0'(id=0) during dp_vdev_register (STA
connect)
	- At this point DP-L0 is linked to two vdevs DP-V0 and DP-V0'
- DP-V0 last ref is released, and it notifier is called
	- DP-L0(id=0) is marked with cdp_vdev_deleted=1
- Interface wlan0 down
	- DP-L0 destroy is called as a part of objmgr vdev destroy
	notification
	- Since DP-L0 has cdp_vdev_deleted=1, it gets freed, even when
	DP-V0' still has the handle of DP-L0 and not deleted.
- DP-V0' gets deleted and calls notifier, but does not find DP-L0 in the
dp_link inactive list and asserts.

In order to mitigate the above scenario, manage a list of cdp_vdev to
which the dp_link has been registered. The dp_link should be freed
only when the list of cdp_vdev is empty. The cdp_vdev is added to the
list during cdp_vdev_register and removed from the list during the
cdp_vdev delete notification.

Change-Id: I39b4e03a7bb3b20736a7a45fcd4210c83208649b
CRs-Fixed: 3948864