qcacld-3.0: Add link switch check for CONNECTING advance logging
CONNECTING advance logs is printed when connecting to assoc link
and partner link which is not expected as the CONNECTING event is
only need to be logged only once when connection is initiated
from the userspace.

Add link switch check for CONNECTING advance logging in order to
prevent CONNECTING advance logs to printed during link
switch.

Add assoc vdev check for CONNECTING advance logging in order to
prevent CONNECTING advance logs to printed for partner link

Change-Id: I4b2a2c02218880fa07196d8db4daa8095b0e7475
CRs-Fixed: 3945391