qcacld-3.0: Delete crypto keys during idle shutdown
Currently, when STA is connected to an AP, driver caches the crypto
keys and after disconnection, it does not delete the crypto keys.
These old crypto keys are deleted on new connection or on the
PSOC close (if driver is not in recovery state) in the stop modules
function.
During idle shutdown, driver calls stop module function and in
between, driver also goes into recovery state (before PSOC close)
while executing the stop module function.
Thus, driver is unable to flush the crypto keys as it is in recovery
state on PSOC close. This leads to memory leak.
Also, this is the race condition between stop module and driver
recovery as in recovery state, stop module does not check for
memory leaks.

So, to fix this,
1. sync up with the driver recovery flag in stop module and PSOC close.
2. delete the STA keys during stop modules

Change-Id: I7ce2c4097a05a0b1df7d060772226a8f03e6d6bf
CRs-Fixed: 3941816