qcacld-3.0: Processes LINK_STATE_CONTROL command for 3 links
With the value of MLD_MAX_SUPPORTED_LINKS set to 2, the host
processes the LINK_STATE_CONTROL command only for the first
2 links present in the following command:
driver ML_LINK_STATE_CONTROL 1 3 1 0 0 1 0.
This command forces the 6 GHz link(link id 3) to be active and
the 2/5 GHz links (link id 0 and 1) to be inactive.

To process the link state control command for all 3 links,
increase the value of MLD_MAX_SUPPORTED_LINKS to 3.

Change-Id: Ib1b3c78c41e0242860c8a665081d58376f871172
CRs-Fixed: 3910353