qcacld-3.0: Use 6 Mbps rate for SAE authentication on 2.4 GHz IoT APs
When performing SAE authentication, connections and roaming to IoT APs
on the 2.4 GHz band were failing because the 6 Mbps rate was not set,
defaulting to the CCK rate instead.

This fix ensures that SAE authentication uses the 6 Mbps rate for both
initial connections and roaming on 2.4 GHz IoT APs.

Change-Id: I2ec57373c07584676ef9ecb049bc668fb7016d6a
CRs-Fixed: 3947657