qcacld-3.0: Handle NDP peer delete during SSR
Currently, driver deletes NDP peer on NDP end indication from the
target. But during SSR, target does not provides the NDP end
indication and driver is unable to delete the NDP peers. Thus, this
results in peer leak when driver checks for memory leaks in stop
modules.

So, to fix this, cleanup the NDP peer in the driver during SSR

Change-Id: Ib3e4dd7ec736e358214085682013385a7bd822a5
CRs-Fixed: 3923353