qcacld-3.0: Set TWT requestor support based on vendor cmd
Currently, Host is only updating the TWT requestor support
in HE caps of mlme config struct based on the vendor command
value. This is not reflected in the association request.
Hence, update this capability in twt psoc object as well to
indicate the same in the association request frames.

CRs-Fixed: 3956407
Change-Id: I9f2673319c3a435519f3670e6e9d5154ce8fe444