qca-wifi: Handle preCAC timer cancel on OCAC completion
When the firmware (FW) sends a delayed OCAC completion status,
the host might have already changed the preCAC channel before
receiving the OCAC completion event. Therefore, the OCAC completion
status should be validated to ensure it corresponds to the currently
configured agile channel. If it does, the preCAC timer must
be cancelled.

In the current code, if the OCAC completion is received on a
channel that is different from the host maintained agile
channel (dfs_agile_precac_freq_mhz), the preCAC timer is
cancelled and hence the next radar detection on the
agile channel fails. This is because
dfs_is_radarsource_agile() considers the radar pulse as agile
radar only if the preCAC timer is running on the host.

To fix this issue, cancel the preCAC timer only if the OCAC
completion is received for the current agile channel.

CRs-Fixed: 3943348
Change-Id: Iba94ddd811b024b9890af076846dbd47efe9e9ae