qcacld-3.0: Wait for roam disable status from FW only for LFR-3.0
The roaming set disable return fail only happened on the
Rome(LFR2.0), and it is just the return fail, but the roaming
disable successfully, that's to say, the roaming function
will not work after after set roaming disable.

In current design, host will wait the fw's event about the roaming
disable success or not after send the roaming disable cmds to fw,
but for the Rome chip, the fw does not have the logic to report the
status event to host after disable roaming complete, so the host wait
timeout and return error.

But this design which wait for the fw's event about the roaming
disable success or not is only designed for LFR3.0, but the the Rome
does not support LFR3.0, just only support the LFR2.0, so it does not
need to wait the status for the roaming disable set.

So add the LFR3.0 check before this wait, and if not LFR3.0, then do
not to wait the fw's event after send the roaming disable to fw.

Change-Id: Iea5bc1eea32974950b242e755eac5615ec112796
CRs-Fixed: 3963975