qca-wifi-oss: DFS events for punctured channel
Issue #1:
Consider a case, when we are moving to channel 60 HT80 with a
puncture pattern of 0x1. (Here channel 52 is punctured). Then the
DFS state of the channel will remain as DFS CAC REQUIRED. This is
because we skip the punctured channels when delivering the DFS
events for punctured subchannels.

But ideally, we have to deliver the DFS event for the punctured
sub-channels based on the state of the puncturing state machine.

This patch writes the core DFS APIs to decide the state of the sub-
channel based of DFS puncturing status

If DFS Puncturing is not configured or if DFS - Puncturing SM is
not running, then deliver the initial event for the punctured
subchannel as well. If any running DFS Puncturing SM includes the
subchannel, then deliver the event based on current state of the
puncturing SM.

Issue #2:
In API 'dfs_deliver_cac_state_events', when delivering the DFS
events for the current channel and previous channel, we skip the
event delivery if the channel is radar. This causes the non-NOL
subchannels to retain their channel state. This is wrong for
non-ETSI regdomains.

Fix this by handling the event delivery for NOL affected
subchannels alone.

Issue #3:
When we move from channel 100 HT80 to channel 104 HT80, we deliver
the DFS events for curchan (104 HT80) first and then for prevchan
(100 HT80). This cause issues as the DFS events sent for curchan
gets overwritten by the DFS events sent for prevchan.
To fix this, deliver the DFS events for prevchan first followed
by the current chan.

Change-Id: I38882cacd0518853b9468b7450cec5c983b7851d