qcacld-3.0: Don't set ML-capability for NDI adapter
Currently, ML-capability is set for an adapter when the interface
is of type STA/SAP and it's a 11be target. NDI interface is
created as STA iface initially through add_virtual_intf() and
the device_mode is changed to NDI(opmode 11) later.
So, NDI interface also gets ML-capability as part of interface
creation. But NDI is not an ML-capable iface and mac_addr
fetching fails due to this flag as it's fetched from an
incorrect place(link_info->link_addr). mac_addr is supposed to
be fetched from adapter->mac_addr for non-ML modes.
Reset the capability when NDI params are configured to
NDI interface.

Change-Id: I8f9b107dc22b20a4f05e2f48dbcd449f01fdb002
CRs-Fixed: 3963619