qcacmn: Avoid configuring filter pass mgmt filter in monitor mode
Currently, during monitor mode bringup, FP management filters are
set along with MO filters for the SW2RXDMA input ring. This is
overwriting the default filter pass configuration of routing
management frames to firmware. This is not a problem in standalone
monitor mode. But, in a STA+Monitor mode use case, this will result
in management packets routed to host driver instead of firmware,
creating issues wrt scan. This is true for both WCN7750 & PEACH
hardware.

Avoid filter pass configuration for beacon and probe response frames
for SW2RXDMA input ring when monitor vdev is present. Since filter
pass configuration is not required for monitor mode, it is okay to
skip it in standalone monitor case as well.

Change-Id: I25d3cd3c9471361c947c4411cbb2f47166410509
CRs-Fixed: 3952132