qcacmn: Remove Locally Generated Scan Entry
Currently, the host driver does not remove locally generated
scan entries during disconnect.
This can lead to reconnections using these entries.
In the case of non-TX MBSSID, there is no mechanism to set the
non-TX MBSSID for the generated scan entry, causing it to send
probe requests during reconnect as the non-TX MBSSID is not set.

This change ensures the removal of locally generated scan entries
during disconnect.

Change-Id: Ie2b60a26b666e1405ea6e1357d777505227685de
CRs-Fixed: 3957457