qcacld-3.0: Enable P2P GO concurrency on TWT
Currently, host send the TWT disable command to the firmware in
case of SCC and MCC concurrencies for STA and SAP.
With new requirement to support TWT in different concurrency (such
as P2P GO + STA MCC), driver do not need to send the TWT disable
command for P2P GO + STA concurrencies.

So, to enable this, check for P2P GO concurrency capability and
P2P GO and STA count and WFD R2 mode.

Change-Id: Id315046579542b9f34f6a97ade9643218e90813b
CRs-Fixed: 3983310