qcacmn: Modify RF test mode scan and connection policy
For RF test mode, when the power type value is set to 1 or 2,
the host checks if a connection is possible in SP and VLP power
modes, respectively. If not, the host checks for a possible
connection in LPI power mode.

With the updated requirements, for RF test mode, when the power
type value is set to 1 or 2, scan and connect to the AP only if
a connection is possible in SP and VLP power modes, respectively.
Additionally, do not connect to the AP if its power type does
not match the RF test mode force power type.

Change-Id: Ic900aa17569a4542362db3ce3341d38379888dd2
CRs-Fixed: 4021160