qcacmn: handle error in filter handling in opt_dp
This change is to handle error in filter handling in
opt_dp.

As per current implementation of opt_dp, we do not
notify timeout to IPA in the case where FW failed
to respond for filter release.
This change is to introduce timeout logic for
release. Wait till 100ms for FW to
respond, if FW responds in give time period
notify success else timeout.

CRs-Fixed: 3982203
Change-Id: Idb6aa80e1a9665f57ba7d8c21de1c205b28e11c2