qcacld-3.0: Use skb prototype to get flow info from socket
Currently, skb's IP protocol was being obtained from the
sock struct of the network buffer where tx flow id will be
stored. The sock struct can be of type IPv6 and contain
IPv4 data as this structure is backward compatible which
is causing some tx flows to get missed.
The packet info is collected from the buffer itself
which internally checks for packet type, which actually
checks network buffers protocol. As the type there will
be absolute, removing the check of sock struct's protocol
type.

Change-Id: I90a5bd06112aab243d3655896ac8efa154d31487
CRs-Fixed: 3933031