qcacld-3.0: Fix BW if SAP hw mode change
In case of SAP hw mode change from eht to he or other mode
bandwidth is not updated and can result in invalid phymode
and start bss failure.

To fix check the max allowed bandwidth and downgrade the
bandwidth if its not supported by the new SAP hw mode.

Change-Id: Iad67ff0f30f6fb7238604a064bad17b5ee77ce86
CRs-Fixed: 3976947