qcacld-3.0: Defer regulatory disconnect until link switch is completed
While the linkswitch is ongoing on vdev-0 regulatory disconnect is triggered,
During this regulatory disconnect posting vdev-1 was the assoc vdev, so disconnect
was posted on vdev-0 to serialization 1st followed by the disconnect on
vdev-1.

Once the linkswitch disconnect is completed, the link switch will be
aborted. After this disconnect will be processed on the vdev0 and
disconnect indication will be sent to userspace which will trigger
the mac address update before vdev1 is properly cleaned up.

So, defer the reulatory disconnect until link switch is completed.

CRs-Fixed: 3958516
Change-Id: Id3729bee01aa5f25010e74e8d6e2a6dd25ce5f83