qcacld-3.0: Update the state of BS state machine correctly
Currently when user space changes the state of bearer switch
state machine to non-wlan, host driver does not update the
state of the BS state machine correctly instead it just update
the state in BS context.
Because of this, state of the BS state machine remains WLAN, which
leads BS state machine to misbehave.

To fix this issue, update state of the BS state machine correctly.

Change-Id: If48e29d1dec59ec959a9dad5658b5ecd59b1abb8
CRs-Fixed: 3982922