qcacmn: check for the eht op before setting the vdev as ml capable
Currently when the vdev is set as the ml vdev only mlie is checked
and it is set as the ml capable vdev. But while intersecting phymode
it gets phymode from the scan module which will check both the ehtop
and multi link element to mark it as the ML capable AP.

This discrepancy between connection manager/lim is causing in sending
invalid information to firmware.

Fix to validate whether the ehtop is present and set the ml capable
cap in vdev accordingly.

CRs-Fixed: 3939575
Change-Id: I0a2edb6b517edad59f3ee556eeb16de1136fa382