qcacmn: Prefer MLMR over MLSR
Scenario:
a) Vendor roam score algorithm is enabled in DBS/non-DBS target.
b) MLO AP is in 2 + 5 + 6 links, and DUT has scan entries of
5 GHz and 6 GHz links only.
c) 5 GHz link is chosen as assoc link, and 6 GHz link will not be
scored, since it is on interfering frequency. And, 2.4 GHz would
not be scored, since scan entry is not available.
d) With both the partners not being scored, the partner sorting
based on best score does not happen and STA connects with 5 GHz
assoc link + first partner 6 GHz, leading to MLSR.

To fix this, if none of the partners are scored, then re-order
the partner links based on the allowed DBS/SBS combinations,
with least preference to the interfering links.

Change-Id: I760fadb27e4821bc727fdc02eb0732e1abd4fdc7
CRs-Fixed: 3984483