qcacmn: Refine candidate list duplicate entry remove API
The existing logic only removes the similar entries if count of
partner links equal, which fails to identify candidates where valid
partner links are same but count is different due to presence of
invalid links.

Create bitmap MAC address of entries in candidate list and remove the
ones matching in the same MLD. The bitmap will be generated from
valid links hence agnostic to number of parnter links which are invalid.

Change-Id: I75cdc0e44ccd78de6b1765f4dba82dd3102a292b
CRs-Fixed: 3985695