qcacld-3.0: Deinit dp trace in wlan shutdown
In suspend/resume and ssr cases, the APIs hdd_wlan_shutdown() and
hdd_wlan_re_init() are used in pair. However, the dp trace deinit
is not called in hdd_wlan_shutdown() while the dp trace init is
called in hdd_wlan_re_init(). This leads to a memory leak for
vzalloc which is used to allocate dp trace tbl.

Change-Id: I225546a2430a06cf91e0d50679c79f68ce9f72fc
CRs-Fixed: 3992291