qcacld-3.0: Send TWT self capability in disconnected state
Currently driver sends self capability and peer cabability
as part of TWT_GET_CAPABILITY request only if sta/p2p_cli
is in connected state.
But framework can fetch and cache self capabilities during
wifi turn on and it will enable TWT APIs based on it.

So to allow this, send TWT self capability in disconnected
state as well.

Change-Id: I03c7b4f3208ebe8098384e877d83a406a4dd8163
CRs-Fixed: 3994971