qcacld-3.0: Rename mlme_operations_bitmap to rso_disabled_bitmap
Currently, host driver sends RSO_STOP to firmware when a
concurrent connection comes and the corresponding requestor bit
is set in mlme_operations_bitmap. This is used to reset the
corresponding bit when concurrent connection is gone and re-enable
RSO when bitmap becomes 0.
So, rename the bitmap to resemble the functionality.

Change-Id: Ie4ed5db490acf044f8b4cf5bc30e62806c81b9e3
CRs-Fixed: 3997073