qcacld-3.0: Add the suspend type on wakeup host event
wma_wow_wakeup_host_event gets called when wakeup happens for
both system suspend and runtime suspend.It is difficult to know whether
device was in system suspend or runtime suspend when wakeup happens.

To address this issue, enhance the log in wma_wow_wakeup_host_event
to know if the device was system suspended or runtime PM suspended.

CRs-Fixed: 3969108
Change-Id: Icf6438d9e87e9aba4411389eb9a96de02dfd6fa5