qcacld-3.0: Get Link reconfig request from userspace
Fetching User Requests: The code adds support to fetch link
reconfiguration requests from userspace. It checks
whether the request is to add or delete a link and
processes the corresponding data.

Command Rejection Conditions: The command will be rejected if:
i.) A link reconfiguration is already in progress
(either from userspace or BTM request).
ii.) Link reconfiguration is not supported by the firmware or driver.
iii.) The virtual device (vdev) is not in a connected state.

Command Handling: Once the data is fetched, the command handler is
invoked to move the link reconfiguration state.

Change-Id: Id1db6ac395481bd11006b992ad2e7b47a08e9b94
CRs-Fixed: 3997698