qcacld-3.0: Fix 5 GHz AP starting with 40 MHz instead of 80 MHz
When the STA primary frequency is 5300, updating the ACS channel frequency
with PCL results in only 5300 being present in the PCL. When sort weight,
the first frequency (5260) in the 80 MHz bonded frequency is incorrectly
ignored for not being in the ACS list. Additionally, 5300 is ignored
because the center frequency difference isn't 30, leaving no 80 MHz ACS
candidate.

To fix it, don't check ACS freq until get combined_weight, if ACS freq
present in bonded freq, can be valid ACS candidate.

Change-Id: Iad63f8a6c8974f507b99bdf14c5b878ddb6ad4f6
CRs-Fixed: 3996174