qcacld-3.0: Downgrade bw to 160 MHz when 320 MHz has invalid freqs
6 GHz AP advertises supported bandwidth as 320 MHz when
corresponding country code is configured to the AP. STA connects
to the AP in 320 MHz though STA has a different country code
which doesn't support some of the sub frequencies in the
AP advertised 320 MHz ch_width.

wlan_reg_chan_band_to_freq() checks and returns 0 if given seg1
channel is not valid as per the given band_mask. But bandwidth
is not downgraded currently.
Downgrade the bw to 160 MHz as 320 MHz can't be a valid
bw for such channel.

Change-Id: I9f70da5440d79910ba965fdcce5c9ae961a54160
CRs-Fixed: 3984412