qcacld-3.0: Avoid multiple roaming triggers for MCC candidates
In non-DBS targets with gWlanMccToSccSwitchMode=6, the driver currently
triggers a roam invoke command to the FW if the current connection
is MCC.
This initial roaming can cause the FW to continue selecting the same
MCC candidate due to a higher score.  Then driver checks the roamed
candidate as MCC and triggers roaming again, leading to
multiple roaming events.

Fix is prevents the driver from triggering roaming if the last roamed
candidate is MCC, thereby avoiding unnecessary multiple roaming
events.

CRs-Fixed: 4000954
Change-Id: Ide4f67cddcf2dd7735a34b7537957d80956a0e39