qca-wifi-oss: Mark PRECAC NOL when processing NOL IE
Issue #1:
Currently, when we process the NOL IE in an RCSA received from
an associated STA, we add the channels to the NOL but do not mark
them as NOL in the PRECAC tree.

This causes inconsistencies between the NOL list and the PRECAC
tree.

Fix this by marking the channels as radar in the PRECAC tree when
adding them to the NOL.

Issue #2:
When each RCSA Action frame is processed, we parse the NOL IE
and add/update the channels to the NOL list. Now, we will also
mark them as pre-CAC NOL in the PRECAC tree.

Consider we are operating on channel 36 HT160. The RE detects
radar in subchannel 52 and sends an RCSA to the root with an NOL
IE corresponding to channel 52.

When processing the first RCSA, we mark the channels as NOL and
also mark them as pre-CAC NOL in the PRECAC tree. However, when
processing the second RCSA for the same subchannel, we attempt to
mark it as NOL in the PRECAC tree again. Since this channel has
already been marked as pre-CAC NOL, we should check if the channel
is already marked as NOL before trying to mark it again. Instead,
we only check if the number of NOL subchannels has reached the
maximum limit. This check fails because the maximum number of
subchannels is 8 for a 160 MHz subtree, 4 for an 80 MHz subtree,
and so on. Currently, the NOL subchannels count is only 1. As a
result, we incorrectly traverse the PRECAC tree and erroneously
increment the number of NOL subchannels for all parent subtrees,
causing inconsistency in the PRECAC node data structures.

Fix this by checking if the frequency is already marked as
pre-CAC NOL before traversing the PRECAC (sub)tree(s).

Change-Id: I52adf84e479cb0441c4db11a11ce6eba1c8ae7db