qcacld-3.0: Fix ACS sorted list print for non 20 MHz freq
During printing ACS sorted list, only freq with weight less than
SAP_ACS_WEIGHT_ADJUSTABLE are printed.

For bandwidth greater than 20 MHz, the score is multiplier of per 20
bandwidth score, eg for 40 MHz its 20 + 20.

So for 40 MHz and above multiply the SAP_ACS_WEIGHT_ADJUSTABLE
with the number of 20 MHz present in the selected bandwidth,
and print the list with score less than that.

Change-Id: I54d4862dc3edca91d341f7056107b64c47ad70b5
CRs-Fixed: 4015358