qcacld-3.0: Add support for add link via user space
Currently, there is no support to add a provided
link from user space because the provided link
might not be present as a link of any scan entry.
This results in a null MLD address, preventing the
link from being added.

This fix provides the necessary support to add the link.

Change-Id: I0719f08ae4b377cfe29bbffd813ff74484f8305c
CRs-Fixed: 4011540