qcacmn: Drop duplicate NLO match event
During a PNO scan while the device is in suspend state,
the firmware sends the NLO match event twice to fulfill
the NLO match -> NLO complete event sequence. There is a
chance that the second NLO match event might arrive after
the NLO complete event, causing the host to set the PNO
match event received flag to true. Consequently, for the
next PNO scan, if the firmware sends an NLO complete event
without an NLO match, the host will still update the PNO
scan results to the kernel.

To prevent this issue, discard the NLO match event if it is
received within 1 second after the NLO complete event.

Change-Id: I8f318bdf3a0dd69657c2e65d2a277d7e9bfdadaa
CRs-Fixed: 4018351