qcacld-3.0: Do not revise the EHT caps for 320 MHz if not set
Do not update the EHT caps for 320 MHz support if the user configured
lower bandwidth for connection. With the user configuration, the
bandwidth and MCS maps are updated and when the EHT caps is revised
for 320 MHz support without checking the original configuration, it
lead to invalid MCS map for 320 MHz, hence the AP rejected connection.

CRs-Fixed: 3976948
Change-Id: I258f2acfc2b6e47fee5c2633a7e59f49e88a8b03