qcacld-3.0: Send unknown avoid reason if reason not specified
There's a mismatch between what the host is sending and
what the firmware (FW) is expecting when the Avoid list
reason is not specified. The host currently sends an
"other reason" code, but the firmware expects no reason
at all.

To fix this, host sends a reason code of 0 when the Avoid
list reason is not specified. This way, the host will send
the same reason code that the firmware expects, ensuring
proper communication between the two.

Change-Id: I5f58262d96447a6a07fa597e669a2d587e0ac647
CRs-Fixed: 4015248