qca-wifi: Configure GAP based on ADFS running status
ADFS precac and Green AP cannot run simulataneously.

So,
1) when precac is running, if the user had enabled green ap, it is not
allowed to be set. Rather the configuration is stored to re-enable
green AP when the precac is not running. Both the states, channel not
being available and the precac is stopped, and, the precac is aborted
is being treated as precac not running and serves as a condition to
enable green AP.

2) when precac is not running, and if the green AP is running, then
precac is allowed to be set. When precac is set, then green AP is
stopped, and the state the precac has interrupted the green AP is
stored. This allows the precac operation to run and when precac is
completed, it is re-enabled again.

This mutual exclusion is only between GAP and Pre-CAC and not with
RCAC.

Change-Id: Ie0e0bfa7511f1cec0a938481f4f137826de723b3
CRs-Fixed: 3962842
(cherry picked from commit 6b5782e10713c06361bf9d8e8f6115e2056cb867)