qcacld-3.0: Restrict 2 Link force active command for single MAC solution
For the single mac solution, the Force link active command
is allowed to process when performed on more than one link
when connected to a 5 + 6 GHz MLO AP irrespective whether the eMLSR
is enabled via ini configuration.

Add check during processing of force link active command
if the command instructs to bring a single link to active
state or two link to active state when ini "enable_emlsr_mode"
is set to '1'.

Change-Id: Ie1a8fe8798b2ff00cf073a210ae8476ab9ef9885
CRs-Fixed: 4022926