qcacmn: Send both EIRP and PSD TPC if AP advertises both
Currently, the host sends both PSD and EIRP TPC to the FW
if the connection power type is SP and the FW has advertised
the capability to process both PSD and EIRP. However, there
is a possibility that these two conditions might be true, but
the AP might have advertised only one of the EIRP or PSD TPE
IE. In such cases, the host will send both EIRP and PSD TPC to
the FW, but one will be empty. Since the FW expects both to be
filled, it will reject the command, and the TPC won't get set.

To address this issue, add a check to send both EIRP and PSD
TPC only if the AP has advertised both.

Change-Id: I60d7cb2e2df0de2ae80e09fb50ab5a392be34957
CRs-Fixed: 4039501