qcacld-3.0: Move SAP to user frequency if STA disconnect
In the 2.4 GHz SAP1 + 5 GHz SAP2 + STA case, if STA connect
to external AP on 5 GHz, and it's operating channel with
NO-IR flag that cause SAP2 has to move to 2.4 GHz for SCC.

Then once STA disconnect, we should move SAP2 to 5 GHz
again. Current logic will not handle such kind of case,
so if there is another SAP working with current SAP in
the same band, we should not skip it and instead to
check if it is possible to switch back to previous
frequency.

Change-Id: I4733b9b3088a2645aaefa03adee7421b50660067
CRs-Fixed: 4030065