qca-wifi: Filter regulatory wireless modes based on frequency
This commit addresses an issue where the regulatory wireless modes are
not properly filtered based on frequency. The problem occurs
specifically in radios that support both 2GHz and 5GHz bands when the
country is set. In this scenario, the frequency is configured as
2437MHz, but the phymode is incorrectly set to 11A, which is a 5GHz
phymode. This leads to finding a valid channel structure even though
such a combination is invalid.

By checking whether the phymode is supported for the given frequency
in the radio, we can prevent invalid combinations. If the phymode is
deemed invalid for the given frequency, the caller has to retry with
auto mode for the given frequency.

Change-Id: I900cab6eed6ef848abe06b8c795e3dd12c2844eb
CRs-Fixed: 3985835