qcacld-3.0: Re-enable roaming on SET LINK response
After START_BSS completes, the host avoids enabling
roaming in the firmware if a link switch is in progress.
It is the host's responsibility to re-enable roaming
after the link switch is completed. Therefore, cache
the request and send RSO_START to the firmware when
SET_LINK is done.

First, reset RSO_START_BSS and set the newly introduced
request ID: RSO_SET_LINK. Upon receiving the SET LINK
response, the host needs to re-enable roaming based on
RSO_SET_LINK.

Change-Id: Iea083233179ee7332f9d104faff06166742d4be9
CRs-Fixed: 4031144