qcacld-3.0: Fix starting index of beacon_filter_extn_table
In the current implementation of the API hdd_add_beacon_filter,
in beacon filter ext IE map incorrectly sets the bit for the IE
elements starting from the index value WLAN_ELEMID_EXTN_ELEM,
instead of the next index.

This misconfiguration indicates to firmware that host is not
interested in changes related to an element IE, and FW does not
forward the updated beacon to host even when there are changes
in that particular IE value.

To resolve this, the starting index of the ext IE table is
adjusted to follow WLAN_ELEMID_EXTN_ELEM.

CRs-Fixed: 4031506
Change-Id: Ie18af8be6c691cd7fdbebe1de7aca283f59832ff