qca-wifi: Change log level from err to debug
There are prints in SAWF API calls exposed to the
NSS, which get called multiple times for every flows,
even for short lived flows. Any prints in this path is
sometime causing excessive logging.

So, changing the logs from err to debug and can be enabled
as required for debug purposes.

Change-Id: I83ce6513556d573a8b113e75069f06237e5d2b0e
CRs-Fixed: 4045642