qcacmn: Fix ipa logging thread shutdown wait time
In some instances, the IPA logging thread failed
to wake up and cancel the wait queue within 100ms.
This results in the shutdown event being set in the
kthread after it had already been destroyed during
the IPA logging deinitialization process.

To resolve this issue, remove the event and flush
the host logs by making the variable local instead
of global.

CRs-Fixed: 4051338
Change-Id: I35c69d1e1285c184f5db2f8dd90709d9b14c495e