qca-wifi: Add support for checking pre-CAC status on punctured chans
Initially, the AP is configured to operate on channel 104 with 160MHz
bandwidth, with preCAC and interCAC features enabled. The interCAC
channel is set to 36 with 80MHz bandwidth. When interCAC is enabled,
the AP immediately operates on the interCAC channel 36 (80MHz) and
performs preCAC on the configured channel 104 (160MHz). Once preCAC is
completed on channel 104 (160MHz), the AP and BSS switch to channel
104 (160MHz). The preCAC channel width then changes to 160MHz, and
preCAC is performed on channel 36 (160MHz).

With the home channel sub-channel marking feature enabled, radar is
detected on the right edge of channel 104 (160MHz), specifically on
channel 128, causing channel 128 to be added to the NOL
(Non-Occupancy List). The AP then switches to the available 160MHz
channel, 36 (160MHz). If the customer wants to switch back to channel
104 (160MHz) by puncturing channel 128 (i.e., 104, 160MHz, 0x80 pp),
this channel change should be allowed.

Currently, once preCAC is done on channel 104 (160MHz) and the channel
changes from 36 (80MHz) interCAC channel to 104 (160MHz) target
channel, the interCAC channel is updated to 104 (160MHz). According to
the interCAC feature, subsequent channel changes go through the
interCAC channel. In the example above, the customer switches from 36
(160MHz) to 104 (160MHz) with pp 0x80 since 128 is in NOL. However,
since the interCAC channel is updated to 104 and 160MHz is not
available, the channel switch happens to 104 (80MHz), which is the
interCAC channel, and it stays there. It remains there because even
after the NOL timeout for channel 128, preCAC cannot be done on
channel 128 as the preCAC width will be 80MHz. Such a channel change
is not possible.

To facilitate the channel change, we need to honor the user's command
to change to a channel that includes a NOL channel by puncturing it.
This requires bypassing the interCAC channel feature, as we do not
support ADFS + Puncturing. By avoiding the interCAC channel, we
prevent switching to the interCAC channel (104, 80MHz) and proceed
directly to the user-configured channel (104, 160MHz, 0x80 pp). This
change is implemented in commit
Ida04ede61c7bc667c34ccc31fbe102a89c29fd32.

Additionally, if preCAC is already done for the target
channel (104, 160MHz, 0x80 pp), the channel change should be seamless
without performing CAC again. However, the current preCAC check does
not account for the puncturing pattern, causing the
channel (104, 160MHz, 0x80 pp) to be incorrectly marked as preCAC not
done due to channel 128 being in NOL. The preCAC check now considers
the puncturing pattern and verifies if the unpunctured channels have
completed preCAC. If so, the channel is marked as preCAC done. This
change is included in this commit.

This patch introduces a function to check if pre-CAC is completed on
punctured channels. It verifies the pre-CAC status for all
sub-channels of a given channel and ensures that pre-CAC is done on
non-punctured sub-channels.

Change-Id: I726f744347bb5161cad90fdfd17874b9835f49bc
CRs-Fixed: 4065588
(cherry picked from commit 8ab071adc694e1136e1cae42616d8087d55ecaf5)