qcacld-3.0: Allow STA to connect on DFS channel in SAP+SAP(DFS) DBS
In the SAP1 + SAP2 (DFS) DBS concurrency test case,
and the scan list is being trimmed to not include
DFS channel. This will result in STA not able to
scan on DFS channel where SAP2 is already operating.

Hence even when STA + SAP DFS SCC INI is set, still
STA will not be allowed to connect.

To resolve this issue:
1. Not exclude the DFS SAP channel in the scan list. And
the change will be brought in
Ia53734e5f35280c9574462d990ef58d6ec17d6fc
2. If the target frequency that STA attempt to connect
is as same as DFS SAP2 and the connection to be allowed
once STA + DFS SAP SCC INI is set.

Change-Id: I405d1126839cc9566a443c6cacf293390d19c32a
CRs-Fixed: 4078612