qca-wifi: Fix RCAC done check
Problem:
When the AP is operating on channel 108 HT20 and the RCAC is running
on channel 112 HT20, after RCAC is done, the AP sends a CSA to move to
108 HT40. As expected, CAC is not performed in this case. When the AP
is operating on channel 112 HT20 and the RCAC is running on channel
108 HT20, after RCAC is done, the AP switches from 112 HT20 to
112 HT40. In this case, CAC is performed again, which is not expected.

Root Cause:
- The algorithm starts from the first index of both sub-channel arrays
and ignores sub-channels that are the same. However, in the case of
switching from 112 HT20 to 112 HT40, the sub-channel arrays are listed
as {112} and {108, 112} respectively. The algorithm incorrectly
decides that it needs to check RCAC done status for both 108 and 112,
leading to unnecessary CAC for channel 112 HT40.

Solution:
Check RCAC done status for all sub-channels in the subtract set
(currentChannel - previousChannel). Calculate the subtract set and
perform RCAC check for all elements as before.

Change-Id: I766ac7d525f1785fdd31ddd3478cce9c432d200b
CRs-Fixed: 4075192
(cherry picked from commit 13cf8ad06c15142f66ece59b933f3103b8d232f9)