qcacmn: Perform self link-switch for MLO roam abort
In the MLO roaming optimization, with partner link
bringup offloaded to host driver, the partner link
will be deleted in the firmware before the start of
roaming.

Therefore, in the case of roam abort, the host driver
has to explicitly bring back the partner vdev based
on the partner link deletion bitmap in the roam event.

To do this, re-use the existing link switch infra
with the following changes:
a) Allow self link switch.
b) Use unique reason code to perform internal disconnect
instead of explicit disconnect.
c) Do not send link switch confirmation command to firmware.

Change-Id: I5b60db976d2c8a83b120607cfa40ab80159cde01
CRs-Fixed: 4101586