qcacmn: Update proper standby link bandwidth
Userspace fails to update the new bandwidth of the
standby link after CSA on the link. This is due to
wrong chandef update for the standby link during
CSA.

During CSA on the standby link, the new bandwidth for
the link is not updated on the link info. Therefore,
the chandef is calculated with new frequency values
and old bandwidth.

To fix this, update the new channel width to the
link info.

CRs-Fixed: 4086617
Change-Id: Ia5df7fa595f37f45ca9b8731deae0ad214f132fd