qcacmn: Fix potential OOB access of pktlog buffer
ppos argument passed to pktlog_read points to the
starting point of the read operation within the
log buffer. This input argument is of type long
long and is used in calculating the offset into
the log buffer whose out type is int. This could
potentially result in an underflow/overflow since
ppos is not sanitized to be within the bounds of
the buffer.

Fix is to sanitize the input argument ppos to be
within the bounds of the allocated log buffer.

Change-Id: Ieaaa9c2509eaed7fcac26ca9d769cda4c7febac1
CRs-Fixed: 4104887