qcacmn: Resolve Coverity Issue for wlan_logging Flag
In the wlan_logging_thread function, the event flag is checked
using test_bit, and then test_and_clear_bit is called.
In second API call causes an error in Coverity testing due to
not checking the return value of this call.

To resolve this, replace the two calls with a single
call to test_and_clear_bit API.

Change-Id: I9dd832ad2ebe5e4818232da78333b946a043a92e
CRs-Fixed: 4112520