qca-wifi-oss: Enhance DFS NOL handling
When an AP is configured with a 320 MHz bandwidth and radar
is detected on a 20 MHz channel, only that specific channel
will be added to the NOL. Due to the puncture map
requirements, adjacent channels are also included to
maintain a valid puncture pattern. The ExTool command
currently displays these adjacent channels as NOL, causing
issues. This update ensures that a DFS NOL event is sent
only when a channel is punctured and in the NOL. Otherwise,
the channel state will be marked as DFS CAC required,
improving clarity.

Change-Id: Ie125eaa37905e56ae49be55fe527cf063324fa2a
CRs-Fixed: 4109596
(cherry picked from commit 63e8edcc2843b71cca253c2a652a17842865364e)