qca-wifi-oss: Check if channel is puncturable
When an AP is operating on a DFS channel (e.g., channel 100) in 11axa
mode, the CAC is bypassed under the assumption that the channel is
punctured and CAC done in the 'dfs_is_precac_done' API.
However, in 11AX mode, there is no puncturing, and the default puncture
pattern is set to 0xFFFF, indicating an invalid puncture pattern.
The PreCAC API 'dfs_is_precac_done' does not verify the validity of this
pattern and assumes it is a valid punctured channel and CAC
done, thereby skipping the CAC

To resolve this issue, implement a sanity check to determine if the
channel is puncturable.

Change-Id: I961645eda7dc7d58a37d5de74d4eee11a0341d60