qcacmn: Fix SW completion path in fig v1
This is a WAR to address the software completion
issue in fig v1. Since, the TQM is not copying the release
source module, host cannot differentiate between HW
and SW completions. To resolve this, FW uses the last
available release reason value (15). Currently
this path is used in QIPS(IPA).

This issue has been resolved by HW in fig V2.

Change-Id: Ie350991f4fe4664d03824eb648aab92bf7f95fcc
CRs-Fixed: 4158498