qcacmn: Add check for wds case in peer map
In current crash we are seeing a scenario in which
peer map event comes for peer id which is already present in
peer_id_to_obj map which results in wds case. This can happen only in
cases:
1. peer unmap not received.
2. peer unmap came with different/invalid peer id.

In wds case if peer is not found with the peer id then ast entry found
with mac address is freed. In current scenario that freed ast entry
is self_ast_entry. This leads to crash in peer teardown
since ast entry is freed and now contains garbage value.

Add a flag wds_not_supported in dp_soc and check for wds case
in peer map if that flag is true.

Change-Id: If21a63910ee4fd400032822b12e1bbc5851f2cb6
CRs-Fixed: 4116508