qcacmn: Set linkid post 1x roam
Scenario:
1) DUT roams using EAPOL Suite-B AKM with 2 links.
2) For this AKM, EAPOL handling is offloaded to the
supplicant.
3) Therefore during roam sync, host driver brings up
the assoc vdev and updates the link id of all the
links in sta_ctx.
4) After EAPOL completion, driver tries to bringup
the partner vdev. During the cleanup of the old
partner vdev, the link-id from sta_ctx is cleared out
for vdev-1.
5) As part of the successful bring up of the new partner
vdev, the sta_ctx should be updated with the link ID.

However, currently the driver updates the new partner
link id only for OWE AKM. This needs to be extended to
all the AKMs which undergo EAPOL offload to supplicant.

To fix this, retain the auth status of the roaming until
the OSIF connection completion update by moving the API
which clears the auth status mlo_roam_connect_complete()
to post OSIF connection update.

Change-Id: I9cc7de69b6c1042c5989a830f3ba490cfc79576d
CRs-Fixed: 4171396