qcacmn: Handling traffic end indication config
When TWT traffic end indication is enabled in dp vdev
and EOT indication is set in service class and flush
indication is set in packet metadata, host creates
a dummy packet as traffic end indication packet and
send it to firmware signifying end of taffic.
Change is disabling LAPB with TWT enabled.

Change-Id: I414bf4deada4dc845d7e6bfef936609e068b77f5
CRs-Fixed: 4207680