qcacmn: Use non-deferrable timer for periodic work
Bus bandwidth periodic work is started after connection
to track the throughput level every 100msecs. The
periodic work is created using a deferrable timer
which does not cause the CPU core to wake up on expiry.
This could lead to unexpected delays in DDR/SNOC voting
due to timer delays which would have impact on throughput
ramp up and stability.

Fix is to use non-deferrable timer for periodic work
which wakes up the CPU core and votes for interconnects
on time.

Change-Id: If82dcecaa207c2dc8e3a61651661191ea42165ca
CRs-Fixed: 4121687