qcacmn: Use right datatype to avoid byte truncation
Driver fetches current timestamp to decide whether to request
cached scan report or use the last available scan report.
This timestamp is 64bit datatype and typecasting it wrongly
to 32bit makes the timestamp value shorter than the
actual value.

The comparison with existing report's timestamp always shows
that the delta is above the threshold and driver sends new
request to FW to fetch the cached scan results.

Ensure right datatype is used to accommodate the full value of
timestamp.

Change-Id: Ie93ed8c7bfc39604c6628191c6583e6ac435f726
CRs-Fixed: 4204839